+++ Practica 2/build.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Archivo de configuración de ANT para compilar ejemplo de ANT.
+Contiene los siguientes comandos:
+* clean - Borra tanto las clases, como el JAR como el JavaDoc generado.
+* build - Compila el código fuente *.java en binarios de Java *.class
+* jar - Compila el código fuente *.java y empaqueta los resultados en un archivo
+	ejecutable *.jar
+* doc - Crea el Javadoc del proyecto.
+* Author: Carlos Cruz Rangel y Juan Carlos Zenteno Pompa.
+* Version: Septiembre 2021. -->
+
+<project>
+
+  <target name="clean" >
+    <delete dir="build" />
+    <delete dir="doc" />
+  </target>
+
+  <!-- Genera todos los archivos .class -->
+  <target name="build" >
+    <mkdir dir="build/classes" />
+    <javac srcdir="src" destdir="build/classes" debug="true" debuglevel="lines,vars,source" >
+    </javac>
+  </target>
+
+  <!-- Generar el archivo ejecutable del proyecto -->
+  <target name="jar" >
+    <antcall target="build" />
+    <mkdir dir="build/jar" />
+    <jar destfile="build/jar/DoubleLinkedList.jar" basedir="build/classes" >
+      <manifest >
+	<attribute name="DependencyFinder" value="main" />
+	<attribute name="Main-Class" value="fciencias.edatos.Practica02.DoubleLinkedList" />
+      </manifest >
+    </jar >
+  </target>
+
+  <!-- Ejecuta el proyecto -->
+  <target name="run" >
+    <java jar="build/jar/DoubleLinkedList.jar" fork="true" >
+    </java>
+  </target >
+
+  <!-- Genera la documentacion -->
+  <target name="doc" >
+    <mkdir dir="doc" />
+    <javadoc sourcepath="src" destdir="doc" access="private" windowtitle="Practica02" doctitle="Practica02" author="true" version="true" >
+      <classpath >
+      </classpath >
+    </javadoc>
+  </target >
+
+</project>
