+++ Practica 2/src/fciencias/edatos/Practica02/TDAList.java
@@ -1,4 +1,6 @@
+package fciencias.edatos.Practica02;
 import java.util.Iterator;
+
 
 /**
 * TDA para listas.
